+++ src/openfermion/resource_estimates/pbc/resources/__init__.py
@@ -18,12 +18,6 @@ except (ImportError, ModuleNotFoundError) as err:
     pytest.skip(f"Need pyscf for PBC resource estimates {err}",
                 allow_module_level=True)
 
-from .compute_lambda_thc import compute_lambda
-from .compute_thc_resources import compute_cost
-from .thc_integrals import (
-    KPTHCDoubleTranslation,
-    KPTHCSingleTranslation,
-)
-from .generate_costing_table_thc import generate_costing_table
-from .factorizations.isdf import solve_kmeans_kpisdf
-from .factorizations.thc_jax import kpoint_thc_via_isdf
+from .data_types import PBCResources, ResourceEstimates
+from .qrom import QR2, QR3, QI2 
+from .rotation_param import compute_beta_for_resources
