+++ src/openfermion/resource_estimates/pbc/resources/data_types_test.py
@@ -12,13 +12,11 @@
 #   limitations under the License.
 import numpy as np
 
-from openfermion.resource_estimates.pbc.utils.resource_utils import (
+from openfermion.resource_estimates.pbc.resources.data_types import (
     PBCResources,
     ResourceEstimates,
-    QR2,
-    QI2,
 )
-from openfermion.resource_estimates.pbc.utils.hamiltonian_utils import (
+from openfermion.resource_estimates.pbc.hamiltonian import (
     HamiltonianProperties,
     build_momentum_transfer_mapping,
 )
@@ -58,33 +56,6 @@ def test_pbc_resources():
     )
     assert (df.toffolis_per_step.values == [919, 366, 895, 787, 949]).all()
     assert (df.total_toffolis.values == [615, 554, 391, 444, 112]).all()
-
-
-def test_momentum_transfer_map():
-    from pyscf.pbc import gto
-
-    cell = gto.Cell()
-    cell.atom = """
-    C 0.000000000000   0.000000000000   0.000000000000
-    C 1.685068664391   1.685068664391   1.685068664391
-    """
-    cell.basis = "gth-szv"
-    cell.pseudo = "gth-pade"
-    cell.a = """
-    0.000000000, 3.370137329, 3.370137329
-    3.370137329, 0.000000000, 3.370137329
-    3.370137329, 3.370137329, 0.000000000"""
-    cell.unit = "B"
-    cell.verbose = 0
-    cell.build(parse_arg=False)
-    kpts = cell.make_kpts([2, 2, 1], scaled_center=[0.1, 0.2, 0.3])
-    mom_map = build_momentum_transfer_mapping(cell, kpts)
-    for i, Q in enumerate(kpts):
-        for j, k1 in enumerate(kpts):
-            k2 = kpts[mom_map[i, j]]
-            test = Q - k1 + k2
-            assert (np.amin(np.abs(test[None, :] - cell.Gv - kpts[0][None, :]))
-                    < 1e-15)
 
 
 def test_qr2():
