+++ src/openfermion/resource_estimates/pbc/sparse/compute_sparse_resources.py
@@ -19,9 +19,8 @@ from numpy.lib.scimath import arccos, arcsin  # has analytc continuation to cplx
 from sympy import factorint
 from openfermion.resource_estimates.utils import QI
 
-from openfermion.resource_estimates.pbc.utils.resource_utils import (
-    ResourceEstimates,)
-from openfermion.resource_estimates.pbc.utils.resource_utils import QR3
+from openfermion.resource_estimates.pbc.resources import (
+    ResourceEstimates, QR3)
 
 
 def compute_cost(
@@ -32,8 +31,7 @@ def compute_cost(
         dE_for_qpe: float = 0.0016,
         chi: int = 10,
 ) -> ResourceEstimates:
-    """Determine fault-tolerant costs using sparse representaion of symmetry
-        adapted integrals.
+    """Determine fault-tolerant costs using sparse representaion of Hamiltonian.
 
     Light wrapper around _compute_cost to automate choice of stps paramter.
 
@@ -86,8 +84,7 @@ def _compute_cost(
         Nky: int,
         Nkz: int,
 ) -> Tuple[int, int, int]:
-    """Determine fault-tolerant costs using sparse representaion of symmetry
-        adapted integrals.
+    """Determine fault-tolerant costs using sparse representaion of Hamiltonian 
 
     Arguments:
         n: the number of spin-orbitals
