+++ src/openfermion/resource_estimates/pbc/thc/compute_thc_resources.py
@@ -18,9 +18,9 @@ from numpy.lib.scimath import arccos, arcsin  # has analytc continuatn to cplx
 from sympy import factorint
 from openfermion.resource_estimates.utils import QI
 
-from openfermion.resource_estimates.pbc.utils.resource_utils import (
+from openfermion.resource_estimates.pbc.resources.data_types import (
     ResourceEstimates,)
-from openfermion.resource_estimates.pbc.utils.resource_utils import QR3
+from openfermion.resource_estimates.pbc.resources.qrom import QR3
 
 
 def compute_cost(
